rename: Server Driven UI 관련 도메인 클래스를 admin 패키지로부터 분리
