+++ packy-api/src/main/java/com/dilly/auth/application/AuthService.java
@@ -7,14 +7,13 @@ import java.util.Optional;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.dilly.admin.ProfileImage;
-import com.dilly.admin.domain.ProfileImageReader;
 import com.dilly.auth.domain.KakaoAccountReader;
 import com.dilly.auth.domain.KakaoAccountWriter;
 import com.dilly.auth.dto.request.SignupRequest;
 import com.dilly.auth.dto.response.SignInResponse;
 import com.dilly.auth.model.KakaoResource;
 import com.dilly.global.exception.NotSupportedException;
+import com.dilly.gift.ProfileImage;
 import com.dilly.global.response.ErrorCode;
 import com.dilly.global.utils.SecurityUtil;
 import com.dilly.jwt.JwtService;
@@ -25,6 +24,7 @@ import com.dilly.member.Provider;
 import com.dilly.member.Status;
 import com.dilly.member.domain.MemberReader;
 import com.dilly.member.domain.MemberWriter;
+import com.dilly.member.domain.ProfileImageReader;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
