+++ packy-api/src/main/java/com/dilly/gift/domain/MessageReader.java
@@ -1,11 +1,14 @@
-package com.dilly.admin.domain;
+package com.dilly.gift.domain;
 
 import java.util.List;
 
 import org.springframework.stereotype.Component;
 
-import com.dilly.admin.dao.MessageRepository;
 import com.dilly.admin.dto.response.ImgResponse;
+import com.dilly.gift.Message;
+import com.dilly.gift.dao.MessageRepository;
+import com.dilly.global.exception.entitynotfound.EntityNotFoundException;
+import com.dilly.global.response.ErrorCode;
 
 import lombok.RequiredArgsConstructor;
 
@@ -23,5 +26,11 @@ public class MessageReader {
 				.build()
 			)
 			.toList();
+	}
+
+	public Message findById(Long id) {
+		return messageRepository.findById(id).orElseThrow(
+			() -> new EntityNotFoundException(ErrorCode.MESSAGE_NOT_FOUND)
+		);
 	}
 }
