+++ packy-api/src/main/java/com/dilly/gift/domain/MusicReader.java
@@ -1,13 +1,13 @@
-package com.dilly.admin.domain;
+package com.dilly.gift.domain;
 
 import java.util.List;
 
 import org.springframework.stereotype.Component;
 
-import com.dilly.admin.Music;
-import com.dilly.admin.MusicHashtag;
-import com.dilly.admin.dao.MusicRepository;
 import com.dilly.admin.dto.response.MusicResponse;
+import com.dilly.gift.Music;
+import com.dilly.gift.MusicHashtag;
+import com.dilly.gift.dao.MusicRepository;
 
 import lombok.RequiredArgsConstructor;
