+++ packy-domain/src/main/java/com/dilly/gift/Box.java
@@ -1,24 +1,21 @@
-package com.dilly.admin;
+package com.dilly.gift;
 
 import static jakarta.persistence.GenerationType.*;
 
 import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
-import jakarta.persistence.ManyToOne;
 import lombok.Getter;
 
 @Entity
 @Getter
-public class Letter {
+public class Box {
 
 	@Id
 	@GeneratedValue(strategy = IDENTITY)
 	private Long id;
 
-	private String content;
+	private String topImgUrl;
 
-	@ManyToOne(fetch = FetchType.LAZY)
-	private LetterPaper letterPaper;
+	private String bottomImgUrl;
 }
