+++ packy-domain/src/main/java/com/dilly/gift/LetterPaper.java
@@ -1,25 +1,21 @@
-package com.dilly.admin;
+package com.dilly.gift;
 
 import static jakarta.persistence.GenerationType.*;
-
-import java.util.List;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
 import lombok.Getter;
 
 @Entity
 @Getter
-public class Music {
+public class LetterPaper {
 
 	@Id
 	@GeneratedValue(strategy = IDENTITY)
 	private Long id;
 
-	private String youtubeUrl;
+	private String envelopeUrl;
 
-	@OneToMany(mappedBy = "music")
-	private List<MusicHashtag> hashtags;
+	private String writingPaperUrl;
 }
