+++ packy-domain/src/main/java/com/dilly/gift/dao/BoxRepository.java
@@ -1,8 +1,8 @@
-package com.dilly.admin.dao;
+package com.dilly.gift.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.dilly.admin.Box;
+import com.dilly.gift.Box;
 
 public interface BoxRepository extends JpaRepository<Box, Long> {
 }
