+++ packy-domain/src/main/java/com/dilly/gift/dao/LetterPaperRepository.java
@@ -1,8 +1,8 @@
-package com.dilly.admin.dao;
+package com.dilly.gift.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.dilly.admin.LetterPaper;
+import com.dilly.gift.LetterPaper;
 
 public interface LetterPaperRepository extends JpaRepository<LetterPaper, Long> {
 }
