+++ packy-domain/src/main/java/com/dilly/gift/dao/LetterRepository.java
@@ -1,8 +1,8 @@
-package com.dilly.admin.dao;
+package com.dilly.gift.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.dilly.admin.Letter;
+import com.dilly.gift.Letter;
 
 public interface LetterRepository extends JpaRepository<Letter, Long> {
 }
