+++ packy-domain/src/main/java/com/dilly/gift/dao/MusicHashtagRepository.java
@@ -1,8 +1,8 @@
-package com.dilly.admin.dao;
+package com.dilly.gift.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.dilly.admin.MusicHashtag;
+import com.dilly.gift.MusicHashtag;
 
 public interface MusicHashtagRepository extends JpaRepository<MusicHashtag, Long> {
 }
