+++ packy-domain/src/main/java/com/dilly/gift/dao/MusicRepository.java
@@ -1,8 +1,8 @@
-package com.dilly.admin.dao;
+package com.dilly.gift.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.dilly.admin.Music;
+import com.dilly.gift.Music;
 
 public interface MusicRepository extends JpaRepository<Music, Long> {
 }
