+++ packy-domain/src/main/java/com/dilly/gift/dao/ProfileImageRepository.java
@@ -1,8 +1,8 @@
-package com.dilly.admin.dao;
+package com.dilly.gift.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.dilly.admin.ProfileImage;
+import com.dilly.gift.ProfileImage;
 
 public interface ProfileImageRepository extends JpaRepository<ProfileImage, Long> {
 }
