+++ packy-domain/src/main/java/com/dilly/member/Member.java
@@ -3,7 +3,7 @@ package com.dilly.member;
 import static jakarta.persistence.FetchType.*;
 import static jakarta.persistence.GenerationType.*;
 
-import com.dilly.admin.ProfileImage;
+import com.dilly.gift.ProfileImage;
 import com.dilly.global.BaseTimeEntity;
 
 import jakarta.persistence.Entity;
